requirements.txt
paho-mqtt==1.3.1
ruamel.yaml==0.15.34